refactored all tests for new ones in a new file check.rb deleted the old ones on main.rb
