+++ spec/logic_spec.rb
@@ -1,19 +1,19 @@
 require_relative '../lib/logic.rb'
+require_relative '../bin/main.rb'
 
 describe Logic do
   let(:logic) { Logic.new }
   describe '#scraper' do
     let(:rand_num) { rand(1..3) }
-    it 'matches output?' do
+    it 'matches output? with valid input' do
       expect(logic.scraper(rand_num)).to eq(logic.final_array)
     end
   end
 
   describe '#scraper' do
     let(:rand_num) { rand(1..3) }
-    it 'matches output?' do
+    it 'matches output? with invalid input' do
       expect(logic.scraper('cxzc')).to eq([])
     end
   end
-
 end
